Address PR feedback except separating stdout and stderr
