+++ provider/pkg/provider/local/base.go
@@ -34,8 +34,6 @@ type BaseInputs struct {
 // Implementing Annotate lets you provide descriptions and default values for fields and they will
 // be visible in the provider's schema and the generated SDKs.
 func (c *BaseInputs) Annotate(a infer.Annotator) {
-	c.CommonInputs.Annotate(a)
-
 	a.Describe(&c.Interpreter, "The program and arguments to run the command.\n"+
 		"On Linux and macOS, defaults to: `[\"/bin/sh\", \"-c\"]`. On Windows, defaults to: `[\"cmd\", \"/C\"]`")
 	a.Describe(&c.Dir, "The directory from which to run the command from. If `dir` does not exist, then\n"+
