+++ provider/pkg/provider/local/commandController_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 
 	"github.com/pulumi/pulumi-command/provider/pkg/provider/common"
-	"github.com/pulumi/pulumi-command/provider/pkg/provider/util"
+	"github.com/pulumi/pulumi-command/provider/pkg/provider/util/testutil"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 	"github.com/stretchr/testify/require"
 )
@@ -35,7 +35,7 @@ func TestOptionalLogging(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			cmd := Command{}
 
-			ctx := util.TestContext{Context: context.Background()}
+			ctx := testutil.TestContext{Context: context.Background()}
 			input := CommandInputs{
 				BaseInputs: BaseInputs{
 					CommonInputs: common.CommonInputs{
