+++ provider/pkg/provider/util/util.go
@@ -16,8 +16,6 @@ package util
 
 import (
 	"bufio"
-	"bytes"
-	"context"
 	"io"
 	"sync"
 
@@ -36,7 +34,8 @@ func LogOutput(ctx p.Context, r io.Reader, doneCh chan<- struct{}, severity diag
 	}
 }
 
-// NoopLogger is for testing. It reads from the provided reader until EOF, discarding the output, then closes the channel.
+// NoopLogger satisfies the expected logger shape but doesn't actually log.
+// It reads from the provided reader until EOF, discarding the output, then closes the channel.
 func NoopLogger(r io.Reader, done chan struct{}) {
 	defer close(done)
 	_, _ = io.Copy(io.Discard, r)
@@ -52,19 +51,3 @@ func (w *ConcurrentWriter) Write(bs []byte) (int, error) {
 	defer w.mu.Unlock()
 	return w.Writer.Write(bs)
 }
-
-// TestContext is a test implementation of p.Context that records all log messages in a buffer, regardless of severity.
-type TestContext struct {
-	context.Context
-	Output bytes.Buffer
-}
-
-func (c *TestContext) log(msg string) {
-	c.Output.WriteString(msg)
-}
-
-func (c *TestContext) Log(_ diag.Severity, msg string)                  { c.log(msg) }
-func (c *TestContext) Logf(_ diag.Severity, msg string, _ ...any)       { c.log(msg) }
-func (c *TestContext) LogStatus(_ diag.Severity, msg string)            { c.log(msg) }
-func (c *TestContext) LogStatusf(_ diag.Severity, msg string, _ ...any) { c.log(msg) }
-func (c *TestContext) RuntimeInformation() p.RunInfo                    { return p.RunInfo{} }
